skip persisted config files when upgrading
add enable and disable scripts to the feature-distributed-locking
to upgrade/downgrade their databases when these features are
enabled or disabled.

allow docker-install.sh and features scripts to override configuration
when the configuration files have been persisted.  These files are
persisted in the context of an OOM installation to save configuration
information across upgrades.

Issue-ID: POLICY-795

+++ packages/docker/src/main/docker/docker-install.sh
@@ -654,7 +654,11 @@ EOF
 	update_monitor $CONTROLLER_NAME
 
 	# save install configuration as an environment file
-	ln -s -f "${POLICY_HOME}/etc/profile.d/${BASE_CONF}" "${POLICY_HOME}/config/${BASE_CONF}.environment"
+	if [[ -f "${POLICY_HOME}/config/${BASE_CONF}.environment" ]] && [[ ! -L "${POLICY_HOME}/config/${BASE_CONF}.environment" ]]; then
+		echo "warning: ${POLICY_HOME}/config/${BASE_CONF}.environment exists as a regular file"
+	else
+		ln -s -f "${POLICY_HOME}/etc/profile.d/${BASE_CONF}" "${POLICY_HOME}/config/${BASE_CONF}.environment"
+	fi
 }
